refactor for "testing"
